WIP: Add forget password feature
This mechanism works similarly to how verification works currently.
The user's email is sent to the server, the server checks if an account
is associated with the email. If so, the user's information will be
stored in a JWT and sent as part of a URL
(i.e., https://www.proctorvue.live/reset-password/<token>) to the email.
Through the link, the user will input their new password, sending it
with the token to the server to reset the password.

In theory, this should work, but I can't test it locally. For some
reason, the path doens't route correctly because of the JWT. The verify
route also doesn't work locally anymore. However, it does work in
production, so I'm assuming this'll work as well.